feat: 新增保存文件功能

+++ crates/text-editor/src/editor/view.rs
@@ -6,6 +6,7 @@ use line::Line;
 use super::{
     editorcommand::{Direction, EditorCommand},
     terminal::{Position, Size, Terminal},
+    DocumentStatus,
 };
 
 mod buffer;
@@ -35,27 +36,27 @@ pub struct View {
     scroll_offset: Position,
 }
 
-impl Default for View {
-    fn default() -> Self {
+impl View {
+    pub fn new(margin_bottom: usize) -> Self {
+        let terminal_size = Terminal::size().unwrap_or_default();
         Self {
             buffer: Buffer::default(),
             needs_redraw: true,
-            size: Terminal::size().unwrap_or_default(),
+            size: Size::from(
+                terminal_size.height.saturating_sub(margin_bottom),
+                terminal_size.width,
+            ),
             text_location: Location::default(),
             scroll_offset: Position::default(),
         }
     }
-}
 
-impl View {
-    pub fn handle_command(&mut self, command: EditorCommand) {
-        match command {
-            EditorCommand::Resize(size) => self.resize(size),
-            EditorCommand::Move(direction) => self.move_text_location(&direction),
-            EditorCommand::Quit => {}
-            EditorCommand::Insert(character) => self.insert_char(character),
-            EditorCommand::Delete => self.delete(),
-            EditorCommand::Backspace => self.backspace(),
+    pub fn get_status(&self) -> DocumentStatus {
+        DocumentStatus {
+            total_line: self.buffer.height(),
+            current_line_index: self.text_location.line_index,
+            file_name: self.buffer.file_name.clone(),
+            is_modified: self.buffer.dirty,
         }
     }
 
@@ -66,15 +67,40 @@ impl View {
         }
     }
 
+    fn save(&mut self) {
+        let _ = self.buffer.save();
+    }
+
+    pub fn handle_command(&mut self, command: EditorCommand) {
+        match command {
+            EditorCommand::Resize(size) => self.resize(size),
+            EditorCommand::Move(direction) => self.move_text_location(direction),
+            EditorCommand::Quit => {}
+            EditorCommand::Insert(character) => self.insert_char(character),
+            EditorCommand::Delete => self.delete(),
+            EditorCommand::Backspace => self.delete_backward(),
+            EditorCommand::Enter => self.insert_newline(),
+            EditorCommand::Save => self.save(),
+        }
+    }
+
     pub fn resize(&mut self, to: Size) {
         self.size = to;
         self.scroll_text_location_into_view();
         self.needs_redraw = true;
     }
 
-    fn backspace(&mut self) {
-        self.move_left();
-        self.delete();
+    fn insert_newline(&mut self) {
+        self.buffer.insert_newline(self.text_location);
+        self.move_text_location(Direction::Right);
+        self.needs_redraw = true;
+    }
+
+    fn delete_backward(&mut self) {
+        if self.text_location.line_index != 0 || self.text_location.grapheme_index != 0 {
+            self.move_text_location(Direction::Left);
+            self.delete();
+        }
     }
 
     fn delete(&mut self) {
@@ -96,7 +122,7 @@ impl View {
             .map_or(0, Line::grapheme_count);
         let grapheme_delta = new_len.saturating_sub(old_len);
         if grapheme_delta > 0 {
-            self.move_right();
+            self.move_text_location(Direction::Right);
         }
         self.needs_redraw = true;
     }
@@ -196,7 +222,7 @@ impl View {
         Position { col, row }
     }
 
-    fn move_text_location(&mut self, direction: &Direction) {
+    fn move_text_location(&mut self, direction: Direction) {
         let Size { height, .. } = self.size;
         match direction {
             Direction::Up => self.move_up(1),
